fix(workspace): unblock CI by satisfying three workspace lints
Main CI was failing three workspace-wide `-D warnings` clippy lints
inside the pre-commit hook, blocking every new commit:

1. `clippy::type-complexity` on
   `CheckerContext::program_wildcard_reexports_type_only` and
   `ProjectEnv::program_wildcard_reexports_type_only` — both typed
   `Option<Arc<FxHashMap<String, Vec<(String, bool)>>>>`. Extract a
   named alias `ProgramWildcardReexportsTypeOnly` alongside the other
   cross-binder aliases.
2. `clippy::useless_conversion` in
   `tsz-core/src/parallel/core.rs:3764` — `Extend::extend` already
   takes `IntoIterator`, drop the redundant `.into_iter()`.
3. Identical `clippy::useless_conversion` at
   `tsz-cli/src/driver/check.rs:2704` (a parallel copy of the same
   helper); fix in lockstep.

Introducing the new alias pushed `context/mod.rs` from 2036 to 2041
lines, over the 2000-line checker-boundary guard. Split the 11
cross-binder/module-resolution `pub type` aliases plus their direct
support types (`ResolutionError`, `ResolutionModeOverride`) into a
sibling `context/aliases.rs` so `mod.rs` lands at 1992 lines.

Repro: https://github.com/mohsen1/tsz/actions/runs/24750230202/job/72411506415
